Refactor: Player implements Observer, simplified TicketSale with Lombok, removed PlayerObserver
