+++ src/main/java/com/model/TicketSale.java
@@ -1,10 +1,12 @@
 package com.model;
 
-import lombok.Getter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
-@Getter
+@Data
+@NoArgsConstructor
+@Builder
 public class TicketSale {
 	private int id;
 	private Player player;
@@ -13,25 +15,17 @@ public class TicketSale {
 	private LocalDateTime saleDate;
 
 	public TicketSale(int id, Player player, Room room, double price, LocalDateTime saleDate) {
-		if (price < 0) throw new IllegalArgumentException("Price must be non-negative");
-		if (player == null || room == null || saleDate == null)
+		if (price < 0) {
+			throw new IllegalArgumentException("Price must be non-negative");
+		}
+		if (player == null || room == null || saleDate == null) {
 			throw new IllegalArgumentException("Player, Room, and saleDate cannot be null");
+		}
 
 		this.id = id;
 		this.player = player;
 		this.room = room;
 		this.price = price;
 		this.saleDate = saleDate;
-	}
-
-    @Override
-	public String toString() {
-		return "TicketSale{" +
-				"id=" + id +
-				", player=" + player +
-				", room=" + room +
-				", price=" + price +
-				", saleDate=" + saleDate +
-				'}';
 	}
 }
