--- src/main/java/com/utils/PlayerObserver.java
@@ -1,10 +1,0 @@
-package com.observerImplementation;
-
-import com.interfaces.Observer;
-
-public class PlayerObserver implements Observer {
-	@Override
-	public void getNotification(String escapeRoomState){
-		System.out.println("Room state is now " + escapeRoomState);
-	}
-}
